Attempt to implement Issue #4 - Add LGDO format conversion utilities
The idea is to add a `convert` function to each LGDO datatype that converts the underlying data to a third-party datatype.
These are `pandas.DataFrame`, `numpy.ndarray` and `awkward.Array`.
Additionally, you have the option to control whether `convert` copies data or not.

At the moment, these issues are still open:

[ ] How to use `to_aoesa` to convert VectorOfVectors to `numpy.ndarray`?
[ ] How to implement the conversion of structures/tables to `numpy.ndarray`?
[ ] How to implement the `convert' function for WaveformTable and encoded data?
[ ] Find out how to implement units with pint. Is it possible for awkward arrays?
[ ] Write many, many tests.